Update encoding / decoding, and rename the old token dataset (see notes)
New encoding is based on formalizing the "3-tuple" structure of encoding. Every note in the midi stream is encoded as 3 values:
-PITCH: a value from 1 to 127 indicating the pitch of the note. 0 means a rest
-DURATION: a value from 0 to 127 indicating the number of 32nd notes in the duration of the note (so max note length is 4 measures)
-ADVANCE: a value from 0 to 127 indicating the number of 32nd notes to advance stream (so max advance is 4 measures)

DURATION and ADVANCE values are quantized to the nearest value if they aren't expressable in 32nd notes (i.e. triplets).

Note that the vocabulary now has a fixed size of 128. This is important for making sure we're generating over the same distribution for each position in the tuple, but it means that we have to use position encodings in the LSTM to distinguish what kind of token we're generating (i.e. PITCH is geneated at position 0,  DURATION at position 1, etc.)